+++ CHANGELOG.md
@@ -258,6 +258,7 @@
 - Manufacturing Order:Unit price problem on consumedStockMoveLineList
 - DECLARATION OF EXCHANGE : stock move lines
 - MANUF ORDER : NPE when no product is selected on producedStockMoveLineList line
+- Disable average price update on cancelling stock move.
 
 ## [Unreleased 5.0.7]
 ## Features